fixed a few issues and axios import. still cannot get my onClick to work. Need to figure out why. Spent all tonight on code challeneg though. I will try to get a TA during work tomorrow.
