+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "5.11.9",
     "@testing-library/react": "11.2.5",
     "@testing-library/user-event": "12.8.0",
+    "axios": "^0.21.1",
     "bookshelf": "^1.2.0",
     "bootstrap": "4.6.0",
     "knex": "^0.21.21",
