+++ src/BestBooks.js
@@ -3,8 +3,23 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import Jumbotron from 'react-bootstrap/Jumbotron';
 import './BestBooks.css';
 import patricktomasso from './patricktomasso.JPEG';
+import { withAuth0 } from "@auth0/auth0-react";
+import axios from 'axios';
 
 class MyFavoriteBooks extends React.Component {
+  // makeAuthReq fixed from App.js in code review
+  makeAuthReq = async () => {
+    const { getIdTokenClaims } = this.props.auth0;
+    let claimThatToken = await getIdTokenClaims();
+    const jwt = claimThatToken.__raw;
+    console.log(jwt);// could console.log the claimThatToken if want to see the long token
+    const config = {
+      headers: { "Authorization": `Bearer ${jwt}` },
+    };
+    const serverResponse = await axios.get('http://localhost:3001/test', config);
+    console.log('I hope it works this way, since it did not work the other way', serverResponse);
+  }
+
   render() {
     return (
       <>
@@ -24,10 +39,12 @@ class MyFavoriteBooks extends React.Component {
           <p>
             This is a collection of my favorite books
           </p>
+          <button onClick={this.makeRequest}>Send me to Server</button>
+          <p>Check the console!</p>
         </Jumbotron>
       </>
     )
   }
 }
 
-export default MyFavoriteBooks;
+export default withAuth0(MyFavoriteBooks);
