+++ src/Login.js
@@ -2,6 +2,7 @@ import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Card from 'react-bootstrap/Card';
 import './Login.css';
+import LoginButton from './LoginButton';
 
 class Login extends React.Component {
   render() {
@@ -12,7 +13,9 @@ class Login extends React.Component {
           <Card.Text>
             Click Below to Log In
           </Card.Text>
-          {/* TODO: add a `LoginButton` component here that will log the user in with Auth0 */}
+          {/* DONE: add a `LoginButton` component here that will log the user in with Auth0 */}
+          <LoginButton />
+
         </Card.Body>
       </Card>
     )
